Python 3 fixes, pymel convert to maya.cmds, and tidy up.
- Switched from tank to sgtk.
- Converted use of pymel to maya.cmds
- Python 3 changes

+++ python/tk_maya/panel_generation.py
@@ -271,7 +271,7 @@ def build_workspace_control_ui(shotgun_panel_name):
 
     # Retrieve the calling Maya workspace control.
     ptr = MQtUtil.getCurrentParent()
-    workspace_control = wrapInstance(long(ptr), QtWidgets.QWidget)
+    workspace_control = wrapInstance(int(ptr), QtWidgets.QWidget)
 
     # Search for the Shotgun app panel widget.
     for widget in QtWidgets.QApplication.allWidgets():
